+++ src/components/signin/Kakao.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { kakaoLogin } from '../../../redux/modules/userSlice';
+import { kakaoLogin } from '../../redux/modules/userSlice';
 
 export default function Kakao() {
   const dispatch = useDispatch();
